requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
distlib==0.3.4
Django==4.0.1
djangorestframework==3.13.1
drf-nested-routers==0.93.4
filelock==3.4.2
install==1.3.5
pipenv==2022.1.8
platformdirs==2.4.1
psycopg2-binary==2.9.3
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
virtualenv==20.13.0
virtualenv-clone==0.5.7
